calculate the accuracy mean as a float

+++ metrics.py
@@ -3,4 +3,4 @@ import torch
 
 def accuracy(predictions, categories):
     _, predicted_categories = torch.max(predictions, 1)
-    return (predicted_categories == categories).sum() / categories.size(0)
+    return (predicted_categories == categories).sum().float() / categories.size(0)
